Exam Preparation

+++ week-10/PeerToPeerChatApp/src/main/java/com/greenfoxacademy/annie/p2pchatapp/Service/AppuserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Service
 public interface AppuserService {
-    Appuser findOne(int id);
+    Appuser findOne(long id);
     List<Appuser> findAll();
     void save(Appuser appuser);
     Appuser findByName(String name);
